Split GameJamController methods from DevNests
